+++ components/Reviews/ReviewsLoop/ReviewsLoop.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Comment } from 'semantic-ui-react'
 
-import ReviewComment from '~/components/Reviews/ReviewComment/ReviewComment'
+import ReviewComment from 'components/Reviews/ReviewComment/ReviewComment'
 
 export default ({ reviews }) => {
     const reviewLoop = reviews ? reviews.map(review => (