+++ pages/shops/menu.js
@@ -1,5 +1,6 @@
 import React from 'react'
 import { connect } from 'react-redux'
+import Router from "next/router";
 import Link from 'next/link'
 import ShopProfile from 'layouts/ShopProfile/ShopProfile'
 import KushyApi from 'utils/KushyApi'
@@ -16,7 +17,7 @@ class ShopMenuPage extends React.Component {
     {
         await api.getProfile('shops', slug)
             .then((results) => (
-                shop = results.data
+                shop = results.data && results.data.length > 0 ? results : Router.redirect('/')
             ))
         await api.getInventory(slug)
             .then((results) => (
@@ -41,7 +42,7 @@ class ShopMenuPage extends React.Component {
       const { shop, inventory, profile } = this.props
       
     return (
-      <ShopProfile shop={ shop } profile={ profile } section="details">
+      <ShopProfile shop={ shop.data[0] } profile={ profile } section="details">
         <section id="menu" className="ui basic segment">
             <h2 className="ui header">
                 <div className="content">